+++ src/app/Components/ComponentsSamples/Labels/LabelsSamples.xaml.cs
@@ -1,0 +1,9 @@
+namespace Components.ComponentsSamples.Labels;
+
+public partial class LabelsSamples
+{
+    public LabelsSamples()
+    {
+        InitializeComponent();
+    }
+}