+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [18.0.0]
+- [BreakingChange] Changed TitleOptions property from FontSize to Style
+- Added styles for Labels
+
 ## [17.26.0] 
 - Resources was updated from DIPS.Mobile.DesignTokens